# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CompletionRequest": {
        "additionalProperties": false,
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "default": "Qwen3 32B Instruct",
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "messages"
        ],
        "title": "CompletionRequest",
        "type": "object"
      },
      "CompletionResponse": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          }
        },
        "required": [
          "message"
        ],
        "title": "CompletionResponse",
        "type": "object"
      },
...
      "DecisionPlaygroundResponse": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/DecisionAnswer"
            },
            "title": "Answers",
            "type": "array"
          },
          "questions": {
            "items": {
              "discriminator": {
                "mapping": {
                  "Choice": "#/components/schemas/ChoiceQuestion",
                  "Noul": "#/components/schemas/NoulQuestion",
                  "Score": "#/components/schemas/ScoreQuestion"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceQuestion"
                },
                {
                  "$ref": "#/components/schemas/ScoreQuestion"
                },
                {
                  "$ref": "#/components/schemas/NoulQuestion"
                }
              ]
            },
            "title": "Questions",
            "type": "array"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "questions",
          "answers"
        ],
        "title": "DecisionPlaygroundResponse",
        "type": "object"
      },
      "DecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "questions": {
            "items": {
              "discriminator": {
                "mapping": {
                  "Choice": "#/components/schemas/ChoiceQuestion",
                  "Noul": "#/components/schemas/NoulQuestion",
                  "Score": "#/components/schemas/ScoreQuestion"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceQuestion"
                },
                {
                  "$ref": "#/components/schemas/ScoreQuestion"
                },
                {
                  "$ref": "#/components/schemas/NoulQuestion"
                }
              ]
            },
            "minItems": 1,
            "title": "Questions",
            "type": "array"
          },
          "state": {
            "minLength": 1,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "questions"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DecisionResponse": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/DecisionAnswer"
            },
            "title": "Answers",
            "type": "array"
          }
        },
        "required": [
          "answers"
        ],
        "title": "DecisionResponse",
        "type": "object"
      },
...
      "GeneratedSpeech": {
        "properties": {
          "meta": {
            "title": "Meta",
            "type": "string"
          },
          "script": {
            "title": "Script",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "voice": {
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "voice",
          "meta",
          "script",
          "time"
        ],
        "title": "GeneratedSpeech",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageResponse": {
        "properties": {
          "image": {
            "$ref": "#/components/schemas/ImageSet"
          }
        },
        "required": [
          "image"
        ],
        "title": "ImageResponse",
        "type": "object"
      },
...
      "RequestResponse": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/RequestRecord"
          }
        },
        "required": [
          "request"
        ],
        "title": "RequestResponse",
        "type": "object"
      },
      "RequestsResponse": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/RequestRecord"
            },
            "title": "Requests",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "requests",
          "total"
        ],
        "title": "RequestsResponse",
        "type": "object"
      },
...
      "SpeechHistoryResponse": {
        "properties": {
          "audio": {
            "items": {
              "$ref": "#/components/schemas/GeneratedSpeech"
            },
            "title": "Audio",
            "type": "array"
          },
          "script": {
            "title": "Script",
            "type": "string"
          },
          "voice_description": {
            "title": "Voice Description",
            "type": "string"
          }
        },
        "required": [
          "audio",
          "voice_description",
          "script"
        ],
        "title": "SpeechHistoryResponse",
        "type": "object"
      },
      "SpeechRequest": {
        "additionalProperties": false,
        "properties": {
          "script": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Script",
            "type": "string"
          },
          "voice": {
            "discriminator": {
              "mapping": {
                "clone": "#/components/schemas/ClonedVoice",
                "describe": "#/components/schemas/DescribedVoice"
              },
              "propertyName": "mode"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/DescribedVoice"
              },
              {
                "$ref": "#/components/schemas/ClonedVoice"
              }
            ],
            "title": "Voice"
          }
        },
        "required": [
          "script",
          "voice"
        ],
        "title": "SpeechRequest",
        "type": "object"
      },
      "SpeechResponse": {
        "properties": {
          "audio": {
            "$ref": "#/components/schemas/GeneratedSpeech"
          }
        },
        "required": [
          "audio"
        ],
        "title": "SpeechResponse",
        "type": "object"
      },
      "TranscriptionRequest": {
        "additionalProperties": false,
        "properties": {
          "audio": {
            "description": "Audio reference; not fetched in mock mode",
            "minLength": 1,
            "title": "Audio",
            "type": "string"
          },
          "formatting": {
            "default": true,
            "title": "Formatting",
            "type": "boolean"
          }
        },
        "required": [
          "audio"
        ],
        "title": "TranscriptionRequest",
        "type": "object"
      },
      "TranscriptionResponse": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TranscriptionResponse",
        "type": "object"
      },
...
      "VideoGenerationRequest": {
        "additionalProperties": false,
        "properties": {
          "aspect": {
            "default": "16:9",
            "enum": [
              "16:9",
              "9:16",
              "1:1"
            ],
            "title": "Aspect",
            "type": "string"
          },
          "duration": {
            "default": 8,
            "exclusiveMinimum": 0.0,
            "title": "Duration",
            "type": "integer"
          },
          "fps": {
            "default": 24,
            "exclusiveMinimum": 0.0,
            "title": "Fps",
            "type": "integer"
          },
          "negative_prompt": {
            "default": "",
            "title": "Negative Prompt",
            "type": "string"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "resolution": {
            "default": "720p",
            "enum": [
              "480p",
              "720p",
              "1080p"
            ],
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "VideoGenerationRequest",
        "type": "object"
      },
      "VideoGenerationResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/VideoJob"
          }
        },
        "required": [
          "job"
        ],
        "title": "VideoGenerationResponse",
        "type": "object"
      },
...
      "VideoResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/VideoJob"
          }
        },
        "required": [
          "job"
        ],
        "title": "VideoResponse",
        "type": "object"
      },
...
      "api__routes__v1__images__edits__ImageRequest": {
        "additionalProperties": false,
        "properties": {
          "aspect": {
            "default": "1:1",
            "enum": [
              "1:1",
              "4:3",
              "3:4",
              "16:9"
            ],
            "title": "Aspect",
            "type": "string"
          },
          "count": {
            "default": 4,
            "enum": [
              1,
              2,
              4
            ],
            "title": "Count",
            "type": "integer"
          },
          "image": {
            "description": "Source image reference; not fetched in mock mode",
            "minLength": 1,
            "title": "Image",
            "type": "string"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "strength": {
            "default": 0.65,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Strength",
            "type": "number"
          }
        },
        "required": [
          "prompt",
          "image"
        ],
        "title": "ImageRequest",
        "type": "object"
      },
      "api__routes__v1__images__generations__ImageRequest": {
        "additionalProperties": false,
        "properties": {
          "aspect": {
            "default": "1:1",
            "enum": [
              "1:1",
              "4:3",
              "3:4",
              "16:9"
            ],
            "title": "Aspect",
            "type": "string"
          },
          "count": {
            "default": 4,
            "enum": [
              1,
              2,
              4
            ],
            "title": "Count",
            "type": "integer"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "ImageRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deterministic mock API matching the frontend fixtures. Inputs are validated, but generation responses are fixed and writes are not persisted. Media is placeholder metadata, not downloadable files. This is not an OpenAI-compatible API.",
    "title": "Kadan API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audio/speech": {
      "get": {
        "operationId": "listSpeech",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Speech",
        "tags": [
          "Audio"
        ]
      },
      "post": {
        "operationId": "generateSpeech",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Speech",
        "tags": [
          "Audio"
        ]
      }
    },
    "/v1/audio/transcriptions": {
      "post": {
        "operationId": "transcribeAudio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcribe Audio",
        "tags": [
          "Audio"
        ]
      }
    },
    "/v1/chat/completions": {
      "post": {
        "operationId": "createCompletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Completion",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/v1/decisions": {
      "get": {
        "operationId": "getDecisions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionPlaygroundResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Decisions",
        "tags": [
          "Decisions"
        ]
      },
      "post": {
        "operationId": "evaluateDecisions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Decisions",
        "tags": [
          "Decisions"
        ]
      }
    },
...
    "/v1/images/edits": {
      "post": {
        "operationId": "editImages",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routes__v1__images__edits__ImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Image",
        "tags": [
          "Images"
        ]
      }
    },
    "/v1/images/generations": {
      "post": {
        "operationId": "generateImages",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routes__v1__images__generations__ImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Image",
        "tags": [
          "Images"
        ]
      }
    },
...
    "/v1/requests": {
      "get": {
        "operationId": "listRequests",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "LLM",
                    "Image",
                    "Video",
                    "TTS",
                    "STT",
                    "Decision"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    200,
                    202,
                    429,
                    500
                  ],
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "title": "Search",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Requests",
        "tags": [
          "Requests"
        ]
      }
    },
    "/v1/requests/{request_id}": {
      "get": {
        "operationId": "getRequest",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Request",
        "tags": [
          "Requests"
        ]
      }
    },
...
    "/v1/videos/generations": {
      "post": {
        "operationId": "generateVideo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Video",
        "tags": [
          "Videos"
        ]
      }
    },
    "/v1/videos/{video_id}": {
      "get": {
        "operationId": "getVideo",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Video",
        "tags": [
          "Videos"
        ]
      }
    }
  }
}